refactor: ziofy ProjectRouteADM (DEV-2104)
* Remove KnoraRoute from ProjectRouteADM
* Replace RouteADMUtil function for random uuid with RouteZ.randomUuid()
* Move factory methods for ProjectUpdatePayloadADM and ProjectCreatePayloadADM into companion object
